LayoutTests/fast/events/touch/ios/netflix-volume-slider-touch-action-quirk-expected.txt

Tests that the Netflix volume-slider quirk prevents page scrolling when dragging the vertical scrubber, but allows scrolling when dragging a horizontal sibling. (rdar://178545839)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS scrollYAfterVerticalScrubberDrag is initialScrollY
PASS scrollYAfterHorizontalScrubberDrag > initialScrollY is true
PASS successfullyParsed is true

TEST COMPLETE
